script/setup: install python packages in venv for docker
Every time the user rebuilds the devcontainer in VSCode, they have to
wait for the installation of all Python dependencies, which takes a
couple of hours on my MacBook M1 and is quite frustrating.

Moreover, newer versions of Python deprecate pip installations without
a virtual environment.

This patch slightly reorganizes the scripts/setup script. Now, the
virtual environment is created in the bind-mounted source
directory. As a result, rebuilding the container will be much quicker,
as pip will be able to reuse previously installed packages.

Signed-off-by: Yuri Volchkov <yuri.volchkov@gmail.com>